+++ src/PSDocs/Commands/CodeCommand.cs
@@ -10,24 +10,27 @@ namespace PSDocs.Commands
     [Cmdlet(VerbsCommon.Format, LanguageKeywords.Code)]
     internal sealed class CodeCommand : KeywordCmdlet
     {
-        private const string ParameterSet_StringDefault = "StringDefault";
+        private const string ParameterSet_Pipeline = "Pipeline";
         private const string ParameterSet_InfoString = "InfoString";
-        private const string ParameterSet_StringInfoString = "StringInfoString";
+        private const string ParameterSet_PipelineInfoString = "PipelineInfoString";
         private const string ParameterSet_Default = "Default";
 
         private List<string> _Content;
 
         [Parameter(Position = 0, Mandatory = true, ParameterSetName = ParameterSet_InfoString)]
-        [Parameter(Position = 0, Mandatory = true, ParameterSetName = ParameterSet_StringInfoString)]
+        [Parameter(Position = 0, Mandatory = true, ParameterSetName = ParameterSet_PipelineInfoString)]
         public string Info { get; set; }
 
-        [Parameter(Position = 0, Mandatory = true, ParameterSetName = ParameterSet_Default, ValueFromPipeline = true)]
-        [Parameter(Position = 1, Mandatory = true, ParameterSetName = ParameterSet_InfoString, ValueFromPipeline = true)]
+        [Parameter(Position = 0, Mandatory = true, ParameterSetName = ParameterSet_Default, ValueFromPipeline = false)]
+        [Parameter(Position = 1, Mandatory = true, ParameterSetName = ParameterSet_InfoString, ValueFromPipeline = false)]
         public ScriptBlock Body { get; set; }
 
-        [Parameter(Mandatory = true, ParameterSetName = ParameterSet_StringDefault, ValueFromPipeline = true)]
-        [Parameter(Mandatory = true, ParameterSetName = ParameterSet_StringInfoString, ValueFromPipeline = true)]
-        public string BodyString { get; set; }
+        [Parameter(Mandatory = true, ParameterSetName = ParameterSet_Pipeline, ValueFromPipeline = true)]
+        [Parameter(Mandatory = true, ParameterSetName = ParameterSet_PipelineInfoString, ValueFromPipeline = true)]
+        [AllowNull]
+        [AllowEmptyString]
+        [AllowEmptyCollection]
+        public object InputObject { get; set; }
 
         protected override void BeginProcessing()
         {
@@ -36,10 +39,8 @@ namespace PSDocs.Commands
 
         protected override void ProcessRecord()
         {
-            if (ParameterSetName == ParameterSet_StringDefault || ParameterSetName == ParameterSet_StringInfoString)
-                AddContent(BodyString);
-            else
-                AddContent(Body.ToString());
+            var content = ParameterSetName == ParameterSet_Pipeline || ParameterSetName == ParameterSet_PipelineInfoString ? InputObject : Body;
+            AddContent(content);
         }
 
         protected override void EndProcessing()
@@ -57,12 +58,13 @@ namespace PSDocs.Commands
             }
         }
 
-        private void AddContent(string input)
+        private void AddContent(object input)
         {
-            if (string.IsNullOrEmpty(input))
+            var s = input?.ToString();
+            if (s == null || string.IsNullOrEmpty(s))
                 return;
 
-            var content = new StringContent(input);
+            var content = new StringContent(s);
             _Content.AddRange(content.ReadLines());
         }
     }