+++ src/PSDocs/Commands/Keyword.cs
@@ -1,5 +1,6 @@
 ﻿
 using PSDocs.Data.Internal;
+using PSDocs.Pipeline;
 using PSDocs.Runtime;
 using System.Management.Automation;
 
@@ -40,19 +41,24 @@ namespace PSDocs.Commands
 
     internal abstract class KeywordCmdlet : PSCmdlet
     {
-        protected ScriptDocumentBuilder Builder
+        protected static ScriptDocumentBuilder Builder
         {
             get { return RunspaceContext.CurrentThread.Builder; }
         }
 
-        protected ScriptDocumentBuilder GetBuilder()
+        protected static ScriptDocumentBuilder GetBuilder()
         {
             return RunspaceContext.CurrentThread.Builder;
         }
 
-        protected PSObject GetTargetObject()
+        protected static PSObject GetTargetObject()
         {
             return RunspaceContext.CurrentThread.TargetObject;
+        }
+
+        protected static PipelineContext GetPipeline()
+        {
+            return RunspaceContext.CurrentThread.Pipeline;
         }
 
         protected static bool True(object o)