+++ src/PSDocs/Common/HashtableExtensions.cs
@@ -1,0 +1,20 @@
+﻿
+using System.Collections;
+using System.Diagnostics;
+
+namespace PSDocs
+{
+    internal static class HashtableExtensions
+    {
+        [DebuggerStepThrough]
+        public static void AddUnique(this Hashtable hashtable, Hashtable values)
+        {
+            if (values == null)
+                return;
+
+            foreach (var key in values.Keys)
+                if (!hashtable.ContainsKey(key))
+                    hashtable.Add(key, values[key]);
+        }
+    }
+}