+++ tests/PSDocs.Tests/TestModule/docs/Test.Doc.ps1
@@ -1,6 +1,6 @@
 
 Document 'TestDocument1' {
-
+    "Culture=$($LocalizedData.Culture)";
 }
 
 Document 'TestDocument2' {