+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+What's changed since pre-release v0.7.0-B2008035:
+
+- Engine features:
+  - Added support for localized strings using the `$LocalizedData` variable. [#91](https://github.com/BernieWhite/PSDocs/issues/91)
+  - Added configuration for setting output options. [#105](https://github.com/BernieWhite/PSDocs/issues/105)
+  - Added parameter alias `-MarkdownEncoding` on `New-PSDocumentOption` for `-Encoding`. [#106](https://github.com/BernieWhite/PSDocs/issues/106)
+- Bug fixes:
+  - Fixed inconsistencies with default options file name. [#103](https://github.com/BernieWhite/PSDocs/issues/103)
+
 ## v0.7.0-B2008035 (pre-release)
 
 What's changed since pre-release v0.7.0-B2008022: